+++ src/Game/Werewolf/Engine.hs
@@ -144,9 +144,9 @@ checkGameOver = do
 
 startGame :: (MonadError [Message] m, MonadWriter [Message] m) => Text -> [Player] -> m Game
 startGame callerName players = do
-    when (playerNames /= nub playerNames)   $ throwError [privateMessage [callerName] "Player names must be unique."]
-    when (length players < 7)               $ throwError [privateMessage [callerName] "Must have at least 7 players."]
-    when (length players > 24)              $ throwError [privateMessage [callerName] "Cannot have more than 24 players."]
+    when (playerNames /= nub playerNames)   $ throwError [privateMessage callerName "Player names must be unique."]
+    when (length players < 7)               $ throwError [privateMessage callerName "Must have at least 7 players."]
+    when (length players > 24)              $ throwError [privateMessage callerName "Cannot have more than 24 players."]
 
     let game = newGame players